requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Deprecated==1.2.10
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyGithub==1.51
PyJWT==1.7.1
pylint==2.5.3
requests==2.24.0
six==1.15.0
smmap==3.0.4
spark-parser==1.8.9
toml==0.10.1
typed-ast==1.4.1
uncompyle6==3.7.1
urllib3==1.25.9
wrapt==1.12.1
xdis==4.7.0
